Cluster script python refactoring
Moving scripts-related code to single class
