+++ cardano_node_tests/tests/test_kes.py
@@ -54,7 +54,9 @@ def short_kes_start_cluster(tmp_path_factory: TempdirFactory) -> Path:
         if destdir_ls:
             return destdir_ls[0]
 
-        startup_files = cluster_nodes.CLUSTER_TYPE.copy_startup_files(destdir=destdir)
+        startup_files = cluster_nodes.CLUSTER_TYPE.cluster_scripts.copy_scripts_files(
+            destdir=destdir
+        )
         with open(startup_files.genesis_spec) as fp_in:
             genesis_spec = json.load(fp_in)
